Restructure the python package.
