+++ python/phylanx/ast/__init__.py
@@ -8,3 +8,5 @@ try:
 
 except Exception:
     from _phylanxd.ast import *
+
+from .transformation import *
